chore(intake): move inline tests to a separate test file
The test module was extracted from `mod.rs` into its own `mod_tests.rs` file to keep the implementation source clean and reduce the size of the main module. The tests themselves are unchanged.

Auto-committed-on: dragonfly
